requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1
